requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-ordered-model==3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Markdown==3.4.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
